update blackjack class, economy class
was getting state issues when I had a dealer instance assigned to the blackjackgame instance. After rearranging so that each call of playJack creates a new dealer instance, seems to work fine.  oh yeah and economy cog is working - need to adapt it's functions to work inside of blackjack cog tho