add queue command
player can leave queue without clearing the whole queue. I want to abstract the player pool/player queue away from the blackjack game in the future. as it is now, blackjack has ownership of the player queue. when implementing future games though, I think it would be easier to work with a separate PlayerQueue class which can be passed as an argument into any game/activity.