add music dependencies to requirements
will be using the google api to search for youtube videos to play, and to store these videos titles + urls. then I can use the youtube-dl package to download whichever URL the user specifies. from there, sending the downloaded audio file to the stream should be pretty easy. I wonder if there's a way to use buffered streams or just inputstreams in order to bypass the need to write the audio to disk. I will work with saving the audio for now, simply to get the feature running - but this idea is something I have been wanting to look into for a while.