music cog changes
for some reason. the music cog throws an error when queuing a new song whil e song is already playing. but it works. it works perfectly if you just ignore that error. so I'm leaving it like this for now, until I figure how to to deal with the timer without throwing exceptions