music cog working perfect
well, more complex is not always better. the way I had the music bot set up before was so convoluted. I'm glad I took a look outside of my own code for how to do this. the music cog is now much easier to undertsand, and works the way its intended to.