skip
bruh. i must not be testing this stuff rigorously enough. anyways I added a boolean to keep track of the state of skipping, to ensure that we don't get a double download situation.
also added a conditional in slugify to make sure both forward and back slashes are not allowed in slugified strings.