started comparing 3 of a kinds, revising 4 of a kind tie case
realized that i can create a getBestKicker() method(recursive perhaps?) that will return the player with the best kicker. this would be useful as many of these tie cases require such an algorithm, and I have been repeating some of these algorithms a lot. plus it would help me fall more in with the dont repeat yourself principle. anyways loggin off for a little to play tarkov but I'll be back to finish this shortly