new music cog
after asking about a solution for my most recent problems with the music cog, Someone suggested I checkout a library called wavelink + lavalink. this is my attempt to implement this new way to stream music to the discord. the cool part about wavelink is that it streams music directly from the internet to your discord voice channel, so there's no intermediate download necessary. of course, there is a trade off in ease of setup as well. the wavelink library requires me to have a lavalink server set up on my network, which is where the music traffic is all routed through.
on the bright side, I think if I go down this route, I may not need to require a google API key from users, or an FFMPEG download for the music feature to work. i;m gonna continue goin gdown this route and see how it turns out. this lilbrary seems very easy to use compared to my complex music cog lmao