basic functionality for MySQLConnection
I realized just now that I still need to figure out how SQL and CSV should handle storing thread_ids. for CSV, i'm not sure if it will be possible without getting real messy. i'll probably have to write another file for storing users guild Ids and thread ids something like this: Username, Guild1Id, Guild2Id, Guild3Id, Guild4Id, GUild5Id. I could store the user's Thread Id for each guild they are a part of + that ParkBot is a part of as values under the GuildXId columns. as for MySQL, I think it will be easier. I can create a new schema for each guild that the bot is a member of, each with userIDs as primary keys for those schema, and user thread Ids as the other value in those rows.