csvconnection complete
all the implementation is done for CSVConnection. bank access + modification is tested + working, and same with threads access + modification.  TODO: I think I still need to cover edge cases for creating the threads file though, I.E. for when there's no data in the csv and we need to write it from scratch.